+++ rig/rig-core/src/providers/copilot/auth.rs
@@ -1,15 +1,24 @@
-use serde::{Deserialize, Serialize};
 use std::fmt;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 
+#[cfg(not(target_family = "wasm"))]
+use serde::{Deserialize, Serialize};
+
+#[cfg(not(target_family = "wasm"))]
 const GITHUB_CLIENT_ID: &str = "Iv1.b507a08c87ecfe98";
+#[cfg(not(target_family = "wasm"))]
 const GITHUB_DEVICE_CODE_URL: &str = "https://github.com/login/device/code";
+#[cfg(not(target_family = "wasm"))]
 const GITHUB_ACCESS_TOKEN_URL: &str = "https://github.com/login/oauth/access_token";
+#[cfg(not(target_family = "wasm"))]
 const GITHUB_API_KEY_URL: &str = "https://api.github.com/copilot_internal/v2/token";
+#[cfg(not(target_family = "wasm"))]
 const DEVICE_CODE_POLL_SLEEP_SECONDS: u64 = 5;
+#[cfg(not(target_family = "wasm"))]
 const DEVICE_CODE_TIMEOUT_SECONDS: u64 = 15 * 60;
+#[cfg(not(target_family = "wasm"))]
 const DEVICE_CODE_SLOW_DOWN_SECONDS: u64 = 5;
 
 #[derive(Debug, Clone)]
@@ -29,6 +38,7 @@ impl DeviceCodeHandler {
         Self(Some(Arc::new(handler)))
     }
 
+    #[cfg(not(target_family = "wasm"))]
     fn emit(&self, prompt: DeviceCodePrompt) {
         if let Some(handler) = &self.0 {
             handler(prompt);
@@ -104,6 +114,7 @@ pub struct AuthContext {
     pub api_base: Option<String>,
 }
 
+#[cfg(not(target_family = "wasm"))]
 #[derive(Debug, Clone, Deserialize, Serialize, Default)]
 struct ApiKeyRecord {
     token: Option<String>,
@@ -111,11 +122,13 @@ struct ApiKeyRecord {
     endpoints: Option<ApiKeyEndpoints>,
 }
 
+#[cfg(not(target_family = "wasm"))]
 #[derive(Debug, Clone, Deserialize, Serialize, Default)]
 struct ApiKeyEndpoints {
     api: Option<String>,
 }
 
+#[cfg(not(target_family = "wasm"))]
 #[derive(Debug, Deserialize)]
 struct DeviceCodeResponse {
     device_code: String,
@@ -125,6 +138,7 @@ struct DeviceCodeResponse {
     expires_in: Option<u64>,
 }
 
+#[cfg(not(target_family = "wasm"))]
 #[derive(Debug, Deserialize)]
 struct AccessTokenResponse {
     access_token: Option<String>,
@@ -132,6 +146,7 @@ struct AccessTokenResponse {
     error_description: Option<String>,
 }
 
+#[cfg(not(target_family = "wasm"))]
 #[derive(Debug, Clone, PartialEq, Eq)]
 struct AccessTokenState {
     token: String,
@@ -401,6 +416,7 @@ impl Authenticator {
     }
 }
 
+#[cfg(not(target_family = "wasm"))]
 impl ApiKeyRecord {
     fn api_base(&self) -> Option<String> {
         self.endpoints
@@ -411,13 +427,14 @@ impl ApiKeyRecord {
 }
 
 #[cfg(not(target_family = "wasm"))]
-fn ensure_parent_dir(path: &Path) -> Result<(), std::io::Error> {
+fn ensure_parent_dir(path: &std::path::Path) -> Result<(), std::io::Error> {
     if let Some(parent) = path.parent() {
         std::fs::create_dir_all(parent)?;
     }
     Ok(())
 }
 
+#[cfg(not(target_family = "wasm"))]
 fn token_expired(expires_at: Option<i64>) -> bool {
     let now = std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
@@ -430,10 +447,12 @@ fn token_expired(expires_at: Option<i64>) -> bool {
     }
 }
 
+#[cfg(not(target_family = "wasm"))]
 fn normalize_poll_interval_seconds(interval: Option<u64>) -> u64 {
     interval.unwrap_or(DEVICE_CODE_POLL_SLEEP_SECONDS).max(1)
 }
 
+#[cfg(not(target_family = "wasm"))]
 fn next_poll_interval_seconds(
     current_interval: u64,
     error: Option<&str>,
@@ -459,6 +478,7 @@ fn next_poll_interval_seconds(
     }
 }
 
+#[cfg(not(target_family = "wasm"))]
 fn format_oauth_error(prefix: &str, error: &str, description: Option<&str>) -> String {
     match description
         .map(str::trim)
@@ -469,6 +489,7 @@ fn format_oauth_error(prefix: &str, error: &str, description: Option<&str>) -> S
     }
 }
 
+#[cfg(not(target_family = "wasm"))]
 fn should_retry_with_fresh_access_token(err: &AuthError) -> bool {
     match err {
         AuthError::Http(err) => should_retry_with_fresh_access_token_status(err.status()),
@@ -476,6 +497,7 @@ fn should_retry_with_fresh_access_token(err: &AuthError) -> bool {
     }
 }
 
+#[cfg(not(target_family = "wasm"))]
 fn should_retry_with_fresh_access_token_status(status: Option<reqwest::StatusCode>) -> bool {
     matches!(
         status,
